UI inconsistencies bug list fixes
