+++ src/components/GenericMenu/GenericMenu.scss
@@ -6,7 +6,6 @@
   @include roboto-bold;
   display: flex;
   justify-content: center;
-  box-shadow: 0 4px 3px -3px #C0C0C0, 0 1px 0 0 #C0C0C0, 0 -1px 0 0 #E0E0E0;
   height: 50px;
   background-color: #fafafb;
   position: relative;
