+++ src/components/RichTextArea/RichTextArea.scss
@@ -18,15 +18,21 @@
 
     .modal-row {
       display: flex;
-      padding: 0 4 * $base-unit;
+      padding: 0 2 * $base-unit;
 
       .object {
         flex-grow: 1;
+        margin-left: 0;
       }
 
       .portrait + .object {
         margin-left: 2 * $base-unit;
       }
+    }
+
+    /* additionally rewrite style with heavier selector */
+    &.modal .modal-row {
+      padding: 0 2 * $base-unit;
     }
 
     .new-post-title {
@@ -120,6 +126,18 @@
         color: $tc-gray-80;
         line-height: $base-unit*5;
         font-size: $tc-label-md;
+        border-color: #aaaaab; /* to match input border colors */
+
+        &:hover {
+          border-color: $tc-gray-50;
+          box-shadow: none;
+        }
+
+        &:focus,
+        &:focus:hover {
+          border-color: $tc-dark-blue-70;
+          box-shadow: 0 0 2px 1px $tc-dark-blue-30;
+        }
       }
 
       &.has-footer {
