+++ src/projects/detail/components/PhaseCard/PhaseCard.scss
@@ -4,7 +4,7 @@
 .phase-card {
   background-color: #fff;
   border-radius: 6px;
-  box-shadow: 0 1px 1px 0 #D5D5D5;
+  box-shadow: 0 1px 1px 0 $tc-gray-10;
 
   @media screen and (max-width: $screen-md - 1px) {
     border-radius: 0;
@@ -56,7 +56,7 @@
   }
   .col {
     &:nth-child(1) {
-      flex: 0 500px;
+      flex: 1 500px;
       @media screen and (max-width: $screen-md - 1px) {
         flex: 1;
         padding-right: 20px;
@@ -67,6 +67,12 @@
     }
     &:nth-child(3) {
       flex: 0 92px;
+    }
+    &:nth-child(4) {
+      display: flex;
+      justify-items: center;
+      flex: 0 55px;
+      min-width: 55px;
     }
     &.show-md {
       display: none;
@@ -254,6 +260,30 @@
           .unit {
             display: none;
           }
+        }
+      }
+    }
+
+    /* to cut shadow on the sides of the specification sections inside project plan phase */
+    .editProjectForm form {
+      overflow: hidden;
+    }
+
+    .dropdown-wrap.default {
+      background-color: $tc-white;
+      border-radius: 4px;
+      border-color: #aaaaab; /* to match color of inputs */
+
+      &:hover {
+        border-color: $tc-gray-50;
+      }
+
+      .dropdown-menu-header {
+        line-height: 18px;
+
+        .tc-link,
+        .tc-link:hover {
+          color: $tc-black;
         }
       }
     }
