+++ src/projects/detail/components/PhaseCardListFooter/PhaseCardListFooter.scss
@@ -1,7 +1,7 @@
 @import '~tc-ui/src/styles/tc-includes';
 
 .container {
-  @include roboto-bold;
+  @include roboto;
   align-items: center;
   background-color: $tc-gray-80;
   border-radius: 0 0 6px 6px;
@@ -26,7 +26,7 @@
 
 .main {
   display: flex;
-  flex: 0 500px;
+  flex: 1 500px;
 }
 
 .total {
@@ -46,7 +46,8 @@
 }
 
 .status {
-  flex: 0 1 40px;
+  flex: 0 55px;
+  min-width: 55px;
 
   @media screen and (max-width: $screen-md - 1px) {
     display: none;
