+++ src/projects/detail/components/PhaseFeed/PhaseFeed.scss
@@ -1,0 +1,3 @@
+.container {
+  margin-top: -1px;
+}
