+++ src/projects/detail/components/PhaseFeed/index.js
@@ -1,0 +1,2 @@
+import PhaseFeed from './PhaseFeed'
+export default PhaseFeed
