+++ src/projects/detail/components/ProjectStageTabs/ProjectStageTabs.jsx
@@ -1,0 +1,31 @@
+import React from 'react'
+
+import GenericMenu from '../../../../components/GenericMenu'
+
+import './ProjectStageTabs.scss'
+
+const ProjectStageTabs = ({
+  activeTab,
+  onTabClick,
+}) => {
+  const tabs = [
+    {
+      onClick: () => onTabClick('posts'),
+      label: 'Posts',
+      isActive: activeTab === 'posts'
+    },
+    {
+      onClick: () => onTabClick('specification'),
+      label: 'Specification',
+      isActive: activeTab === 'specification'
+    }
+  ]
+
+  return (
+    <div styleName="container">
+      <GenericMenu navLinks={tabs} />
+    </div>
+  )
+}
+
+export default ProjectStageTabs
