+++ src/projects/detail/components/ProjectStageTabs/ProjectStageTabs.scss
@@ -1,0 +1,6 @@
+@import '~tc-ui/src/styles/tc-includes';
+
+.container {
+  border-top: 1px solid $tc-gray-10;
+  border-bottom: 1px solid $tc-gray-10;
+}
