+++ src/projects/detail/components/ProjectStageTabs/index.js
@@ -1,0 +1,2 @@
+import ProjectStageTabs from './ProjectStageTabs'
+export default ProjectStageTabs
