+++ src/projects/detail/components/TwoColsLayout/TwoColsLayout.scss
@@ -12,7 +12,7 @@
   flex: 1;
   max-width: 360px;
   background-color: $tc-white;
-  box-shadow: 1px 1px 0 0 #DCDCE0, 0 1px 0 0 #DCDCE0;
+  box-shadow: 1px 1px 0 0 $tc-gray-10, 0 1px 0 0 $tc-gray-10;
   min-height: calc(100vh - 110px);
   padding-bottom: 10 * $base-unit;
   position: relative;
