+++ src/projects/detail/containers/FeedContainer.scss
@@ -1,5 +1,11 @@
 @import '~tc-ui/src/styles/tc-includes';
+@import '../../../styles/includes';
 
 .chat-button-space {
   height: 14 * $base-unit;
+}
+
+.feed-card {
+  @include stand-alone-item-shadow;
+  border-radius: $card-border-radius;
 }
